Bug fix for MRI's parser bug
use \xAA\xAA instead of \uAAAA. \u triggers bug in Ruby parser
to turn binary string into UTF8.
force encoding of Lexer/Parser files to binary, use .bytes to
split the string, then force two byte reads per character
and handle the numerical shift in the Antlr encoding in the readInt
to return a two-byte character from the string.

The version is not shifted so special instructions are used to handle
that case.